Build new resource IDs into config.js
This change passes the IDs and URLs of important new resources
(AdminUserPoolId, AdminUserPoolClientId, and AdminUserPoolDomain) to
the static-asset-uploader, where they're built into config.js for use
on the front-end.

This change also removes all references to the UseCognitoHostedUI
parameter; HostedUI will be mandatory moving forward, and the flag will
no longer be provideable.