Prevent users from accessing S3 directly using OAI
This change configures an origin access identity (OAI) on the CloudFront
distribution and adds a policy to the S3 bucket allowing only that OAI
to access the bucket contents. This means that users can no longer
directly reach S3; all traffic will go through CF.

This change also changes the default ACL on the files uploaded by the
static-asset-uploader, because those appear to take precedence?