Modify frontend UI to match new auth
The UI used to have to detect if the dev portal had been set up for
hosted or unhosted sign-in; since hosted sign-in is now mandatory, that
logic is torn out. Additionally, users now have to specify if they're
logging in as a user or admin, since they redirect to different user
pool domains.