Add custom resource for resource servers
Note that it uses util.inspect() instead of JSON.stringify() and
returns a static message instead of responseData. This is because some
part of responseData is circular, and throws when stringified. This is
just a short-term work-around; we should make a more permanent fix and
port it back into the other 2 notify-cfn.js files floating around.