Move to promisified notifyCFN
If we return a promise that doesn't fully capture the work done in
non-promisified notifyCFN, the lambda runtime exits too early. If we
don't return a promise, the tests don't pass. This change moves to
promifisied notifyCFN so that the runtime survives until all work is
done and the tests pass.